+++ app/javascript/src/NewApplication/index.js
@@ -6,8 +6,6 @@ export * from './components/ProductSelect'
 export * from './components/ApplicationPlanSelect'
 export * from './components/NameInput'
 export * from './components/DescriptionInput'
-export * from './components/SelectProductModal'
-export * from './components/SelectBuyerModal'
 // $FlowFixMe $FlowError[cannot-resolve-module] failing because file name contains module's name NewApplication
 export { NewApplicationForm } from './components/NewApplicationForm'
 // $FlowFixMe $FlowError[cannot-resolve-module] failing because file name contains module's name NewApplication
